+++ webknossos/webknossos/client/api_client/models.py
@@ -151,6 +151,11 @@ class ApiDatasetUploadInformation:
 
 
 @attr.s(auto_attribs=True)
+class ApiDatasetUploadSuccess:
+    new_dataset_id: str
+
+
+@attr.s(auto_attribs=True)
 class ApiLinkedLayerIdentifier:
     organization_id: str
     data_set_name: str
@@ -164,6 +169,7 @@ class ApiReserveDatasetUploadInformation:
     name: str
     organization: str
     total_file_count: int
+    total_file_size_in_bytes: int
     initial_teams: List[str]
     layers_to_link: Optional[List[ApiLinkedLayerIdentifier]] = None
     folder_id: Optional[str] = None