+++ cluster_tools/Changelog.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/) `MAJOR.MIN
 For upgrade instructions, please check the respective *Breaking Changes* sections.
 
 ## Unreleased
-[Commits](https://github.com/scalableminds/webknossos-libs/compare/v2.0.4...HEAD)
+[Commits](https://github.com/scalableminds/webknossos-libs/compare/v2.0.7...HEAD)
 
 ### Breaking Changes
 
@@ -16,6 +16,18 @@ For upgrade instructions, please check the respective *Breaking Changes* section
 ### Changed
 
 ### Fixed
+
+
+## [2.0.7](https://github.com/scalableminds/webknossos-libs/releases/tag/v2.0.7) - 2025-03-26
+[Commits](https://github.com/scalableminds/webknossos-libs/compare/v2.0.6...v2.0.7)
+
+
+## [2.0.6](https://github.com/scalableminds/webknossos-libs/releases/tag/v2.0.6) - 2025-03-25
+[Commits](https://github.com/scalableminds/webknossos-libs/compare/v2.0.5...v2.0.6)
+
+
+## [2.0.5](https://github.com/scalableminds/webknossos-libs/releases/tag/v2.0.5) - 2025-03-25
+[Commits](https://github.com/scalableminds/webknossos-libs/compare/v2.0.4...v2.0.5)
 
 
 ## [2.0.4](https://github.com/scalableminds/webknossos-libs/releases/tag/v2.0.4) - 2025-03-19