+++ webknossos/Changelog.md
@@ -10,16 +10,41 @@ and this project adheres to [Semantic Versioning](http://semver.org/) `MAJOR.MIN
 For upgrade instructions, please check the respective _Breaking Changes_ sections.
 
 ## Unreleased
-[Commits](https://github.com/scalableminds/webknossos-libs/compare/v2.0.4...HEAD)
+[Commits](https://github.com/scalableminds/webknossos-libs/compare/v2.0.7...HEAD)
 
 ### Breaking Changes
 
 ### Added
+- Adds the field `totalFileSizeInBytes` to the dataset upload information. [#1264](https://github.com/scalableminds/webknossos-libs/pull/1264)
+- Added support for metadata in segment and trees. [#1271](https://github.com/scalableminds/webknossos-libs/pull/1271)
+- Added a warning on import if a newer version of the `webknossos` package is available from PyPI. [#1280](https://github.com/scalableminds/webknossos-libs/pull/1280)
 
 ### Changed
 - Using `x-auth-header` to send tokens to datastore. [#1270](https://github.com/scalableminds/webknossos-libs/pull/1270)
 
 ### Fixed
+
+
+## [2.0.7](https://github.com/scalableminds/webknossos-libs/releases/tag/v2.0.7) - 2025-03-26
+[Commits](https://github.com/scalableminds/webknossos-libs/compare/v2.0.6...v2.0.7)
+
+### Changed
+- Pinned `tensorstore` to `<=0.1.71`. [#1276](https://github.com/scalableminds/webknossos-libs/pull/1276)
+
+
+## [2.0.6](https://github.com/scalableminds/webknossos-libs/releases/tag/v2.0.6) - 2025-03-25
+[Commits](https://github.com/scalableminds/webknossos-libs/compare/v2.0.5...v2.0.6)
+
+### Fixed
+- Fixed an issue with uploading datasets where the dataset name is already in use. [#1275](https://github.com/scalableminds/webknossos-libs/pull/1275)
+
+
+
+## [2.0.5](https://github.com/scalableminds/webknossos-libs/releases/tag/v2.0.5) - 2025-03-25
+[Commits](https://github.com/scalableminds/webknossos-libs/compare/v2.0.4...v2.0.5)
+
+### Changed
+- Changed the default dimension separator for Zarr 2 arrays from `.` to `/`. [#1274](https://github.com/scalableminds/webknossos-libs/pull/1274)
 
 
 ## [2.0.4](https://github.com/scalableminds/webknossos-libs/releases/tag/v2.0.4) - 2025-03-19